Added order and task managers
Order and task managers manage the orders assign to the level. The order manager display orders and send order details to task manager, which determine the possible tasks to the sous chefs, as well as display the tasks where the play can choose which to assign to the sous chefs.