Updates to order distributions and prefabs
The orders distribution for the sous chefs now works as the sous chefs are directly producing the doughs based on a single order list for the level (this can be improve up on by create different random lists for each chef).

Also, custom prefabs are created for the pizza models, which include adding rigidbody component, activate convex for mesh collider, and use the correct material for mesh renderer.

And, service table is currently being work on.